requirements.txt
opencv-python
mediapipe
numpy
pyautogui
tensorflow
keras
pyvirtualcam
pillow
